removed unneeded component
